virtual-interviewer-main/pyproject.toml
[project]
name = "virtual-interviewer-git-recovered"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-genai>=1.56.0",
    "google-generativeai>=0.7.2",
    "groq>=1.0.0",
    "librosa>=0.11.0",
    "mediapipe==0.10.9",
    "moviepy>=2.2.1",
    "nltk>=3.9.2",
    "numpy<2.0",
    "openai-whisper>=20250625",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "praat-parselmouth>=0.4.7",
    "protobuf>=3.20.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "gtts",
    "pdfplumber",
    "pypdf",
    "python-docx",
    "striprtf",
    "matplotlib",
    "reportlab",
    "pillow",
    "numpy"
]
